Revert changes of SimpleButton class, added a new MenuButton class, clean up obsolete code
