+++ core/src/com/oniz/Game/AssetLoader.java
@@ -74,13 +74,13 @@ public final class AssetLoader {
         sprites.put("pauseUp", new TextureRegion(textures.get("menuTexture"), 12, 814, 72, 62));
         sprites.put("pauseDown", new TextureRegion(textures.get("menuTexture"), 12, 894, 72, 62));
 
-//        // play, pause, restart buttons
-//        textures.put("pauseUp", new Texture(Gdx.files.internal("lineLight/lineLight12.png")));
-//        textures.put("pauseDown", new Texture(Gdx.files.internal("shadedDark/shadedDark14.png")));
-//        textures.put("playUp", new Texture(Gdx.files.internal("lineLight/lineLight14.png")));
-//        textures.put("playDown", new Texture(Gdx.files.internal("shadedDark/shadedDark16.png")));
-//        textures.put("restartUp", new Texture(Gdx.files.internal("lineLight/lineLight10.png")));
-//        textures.put("restartDown", new Texture(Gdx.files.internal("shadedDark/shadedDark12.png")));
+        // play, pause, restart buttons
+        textures.put("pauseUp", new Texture(Gdx.files.internal("lineLight/lineLight12.png")));
+        textures.put("pauseDown", new Texture(Gdx.files.internal("shadedDark/shadedDark14.png")));
+        textures.put("playUp", new Texture(Gdx.files.internal("lineLight/lineLight14.png")));
+        textures.put("playDown", new Texture(Gdx.files.internal("shadedDark/shadedDark16.png")));
+        textures.put("restartUp", new Texture(Gdx.files.internal("lineLight/lineLight10.png")));
+        textures.put("restartDown", new Texture(Gdx.files.internal("shadedDark/shadedDark12.png")));
 
         // temporary buttons: A, B, X, Y
         textures.put("buttonAUp", new Texture(Gdx.files.internal("lineLight/lineLight34.png")));
