+++ core/src/com/oniz/Screens/GameScreen.java
@@ -1,18 +1,19 @@
 package com.oniz.Screens;
 
-
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.oniz.Game.GameRenderer;
 import com.oniz.Game.GameWorld;
 import com.oniz.Game.InputHandler;
 
+/**
+ * GameScreen class which instantiates and coordinates GameWorld and GameRenderer.
+ */
 public class GameScreen implements Screen {
 
     private GameWorld gameWorld;
     private GameRenderer gameRenderer;
     private float runTime;
-
 
     public GameScreen() {
         float screenWidth = Gdx.graphics.getWidth();
