+++ core/src/com/oniz/Screens/MainScreen.java
@@ -37,10 +37,6 @@ public class MainScreen implements Screen {
 
     private float runTime;
 
-    // temporary buttons to simulate gestures
-    private SimpleButton buttonA, buttonB, buttonX, buttonY;
-
-    private SimpleButton pauseButton, playButton,restartButton;
     private Stage stage;
     private Table table;
 
@@ -53,12 +49,6 @@ public class MainScreen implements Screen {
     public MainScreen(ZGame zGame) {
 
         this.zGame = zGame;
-
-        // temporary
-//        buttonA = new SimpleButton(new Image(AssetLoader.getInstance().textures.get("buttonAUp")).getDrawable(), new Image(AssetLoader.getInstance().textures.get("buttonADown")).getDrawable());
-//        buttonB = new SimpleButton(new Image(AssetLoader.getInstance().textures.get("buttonBUp")).getDrawable(), new Image(AssetLoader.getInstance().textures.get("buttonBDown")).getDrawable());
-//        buttonX = new SimpleButton(new Image(AssetLoader.getInstance().textures.get("buttonXUp")).getDrawable(), new Image(AssetLoader.getInstance().textures.get("buttonXDown")).getDrawable());
-//        buttonY = new SimpleButton(new Image(AssetLoader.getInstance().textures.get("buttonYUp")).getDrawable(), new Image(AssetLoader.getInstance().textures.get("buttonYDown")).getDrawable());
 
         stage = new Stage(new FitViewport(450, 800));
 
@@ -73,39 +63,6 @@ public class MainScreen implements Screen {
         im.addProcessor(gd);
         im.addProcessor(stage);
 
-//        // temporary
-//        buttonA.addListener(new ChangeListener() {
-//            @Override
-//            public void changed(ChangeEvent event, Actor actor) {
-//                System.out.println("Button A Pressed");
-//                gameWorld.killZombie(0);
-//            }
-//        });
-//
-//        buttonB.addListener(new ChangeListener() {
-//            @Override
-//            public void changed(ChangeEvent event, Actor actor) {
-//                System.out.println("Button B Pressed");
-//                gameWorld.killZombie(1);
-//            }
-//        });
-//
-//        buttonX.addListener(new ChangeListener() {
-//            @Override
-//            public void changed(ChangeEvent event, Actor actor) {
-//                System.out.println("Button X Pressed");
-//                gameWorld.killZombie(2);
-//            }
-//        });
-//
-//        buttonY.addListener(new ChangeListener() {
-//            @Override
-//            public void changed(ChangeEvent event, Actor actor) {
-//                System.out.println("Button Y Pressed");
-//                gameWorld.killZombie(3);
-//            }
-//        });
-
         table = new Table();
         table.setFillParent(true);
         table.setDebug(true);
@@ -113,18 +70,7 @@ public class MainScreen implements Screen {
         table.align(Align.top | Align.right);
         table.pad(5);
 
-        // temporary
-//        table.add(buttonA);
-//        table.add(buttonB);
-//        table.add(buttonX);
-//        table.add(buttonY);
-//
-//        table.add(pauseButton);
-//        table.add(playButton);
-//        table.add(restartButton);
         stage.addActor(table);
-
-
     }
 
 
@@ -139,7 +85,6 @@ public class MainScreen implements Screen {
         runTime += delta;
         gameWorld.update(delta);
         gameRenderer.render(runTime);
-
 
 //      draw stage over everything else
         stage.draw();
